accessible/tests/mochitest/tree/test_iframe.html

<!DOCTYPE html>
<html>
<head>
  <title>Outer document accessible tests</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      var accTree = {
        role: ROLE_INTERNAL_FRAME,
        children: [
          {
            role: ROLE_DOCUMENT
          }
        ]
      };

      testAccessibleTree("iframe", accTree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="Fix O(n^2) access to all the children of a container"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=342045">
    Mozilla Bug 342045
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <iframe id="iframe" src="about:mozilla">

</body>
</html>
